refactor wechat sender
